+++ R/array-class.R
@@ -6,14 +6,19 @@ detectFronts.array <- function(x, method = "BelkinOReilly2009", intermediate = F
   # Check arguments
   checkArgs_df_array(x = x)
 
-  output <- if(intermediate) list() else x
+  # output <- if(intermediate) list() else x
   for(i in seq(dim(x)[3])){
     # Apply the basic function
-    tempOut <- detectFronts(x = x[,,i], method = method, intermediate = intermediate, ...)
+    tempOut <- detectFronts(x = x[,,i], method = method, intermediate = intermediate,
+                            checkPrevs = FALSE, ...)
 
     # Depending on 'intermediate', the output will be a single array or a list of them
     if(intermediate){
-      output[[i]] <- tempOut
+      if(i == 1){
+        output <- array(data = NA, dim = c(dim(x), dim(tempOut)[3]))
+      }
+
+      output[,,i,] <- tempOut
     }else{
       output[,,i] <- tempOut
     }